+++ src/enums/boolean.js
@@ -1,0 +1,4 @@
+export const BoolEnums = {
+    bTrue: '1',
+    bFalse: '0'
+};