+++ src/enums/index.js
@@ -3,3 +3,4 @@ export { KeyboardKeys } from './keyboard';
 export { GeneralNoteEnums, GeneralNoteCodes } from './generalNote';
 export { Cursors } from './styles';
 export { ArrayEnums } from './arrays';
+export { BoolEnums } from './boolean';