+++ src/enums/index.js
@@ -1,6 +1,6 @@
 
 export { KeyboardKeys } from './keyboard';
-export { GeneralNoteEnums, GeneralNoteCodes } from './generalNote';
+export { GeneralNoteEnums, GeneralNoteCodes } from '../components/ui/generalNote/enums';
 export { Cursors } from './styles';
 export { ArrayEnums } from './arrays';
 export { BoolEnums } from './boolean';