+++ Changelog.md
@@ -1,6 +1,6 @@
 ## Changelog
 v 2.38
-  Added GameViewer improvements, ContentSwitcher improvements
+  Added GameViewer improvements, ContentSwitcher improvements, added gameViewer theme
 
 v 2.37
   Created GameViewer and ContentSwitcher to view games